CDPSession docs
