+++ lib/PuppeteerSharp.Tests/TargetTests/TargetTests.cs
@@ -18,6 +18,7 @@ namespace PuppeteerSharp.Tests.TargetTests
                 && target.Url == TestConstants.AboutBlank);
             Assert.Contains(targets, target => target.Type == "other"
                 && target.Url == string.Empty);
+
         }
 
         [Fact]
