+++ lib/PuppeteerSharp/Input/Keyboard.cs
@@ -10,12 +10,12 @@ namespace PuppeteerSharp.Input
     /// </summary>
     public class Keyboard
     {
-        private readonly Session _client;
+        private readonly CDPSession _client;
         private readonly HashSet<string> _pressedKeys = new HashSet<string>();
 
         internal int Modifiers { get; set; }
 
-        public Keyboard(Session client)
+        public Keyboard(CDPSession client)
         {
             _client = client;
         }
