+++ lib/PuppeteerSharp/Input/Mouse.cs
@@ -8,7 +8,7 @@ namespace PuppeteerSharp.Input
     /// </summary>
     public class Mouse
     {
-        private readonly Session _client;
+        private readonly CDPSession _client;
         private readonly Keyboard _keyboard;
 
         private decimal _x = 0;
@@ -20,7 +20,7 @@ namespace PuppeteerSharp.Input
         /// </summary>
         /// <param name="client">The client</param>
         /// <param name="keyboard">The keyboard</param>
-        public Mouse(Session client, Keyboard keyboard)
+        public Mouse(CDPSession client, Keyboard keyboard)
         {
             _client = client;
             _keyboard = keyboard;
