+++ lib/PuppeteerSharp/PageCoverage/JSCoverage.cs
@@ -8,14 +8,14 @@ namespace PuppeteerSharp.PageCoverage
 {
     internal class JSCoverage
     {
-        private readonly Session _client;
+        private readonly CDPSession _client;
         private readonly Dictionary<string, string> _scriptURLs;
         private readonly Dictionary<string, string> _scriptSources;
 
         private bool _enabled;
         private bool _resetOnNavigation;
 
-        public JSCoverage(Session client)
+        public JSCoverage(CDPSession client)
         {
             _client = client;
             _enabled = false;
